purge fix

+++ modules/purge.py
@@ -31,11 +31,13 @@ async def del_msg(_, message: Message):
 @Client.on_message(filters.command("purge", prefix) & filters.me)
 @with_reply
 async def purge(client: Client, message: Message):
-    chunk = [message.id]
+    chunk = []
     async for msg in client.get_chat_history(
         chat_id=message.chat.id,
         limit=message.id - message.reply_to_message.id + 1,
     ):
+        if msg.id < message.reply_to_message.id:
+            break
         chunk.append(msg.id)
         if len(chunk) >= 100:
             await client.delete_messages(message.chat.id, chunk)
